changed argument type to int

+++ flarED-H_parser.py
@@ -8,7 +8,7 @@ from Range import Range
 PARSER = argparse.ArgumentParser(description="",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 
-PARSER.add_argument("-he", "--height", type=float, default=None, required=True,
+PARSER.add_argument("-he", "--height", type=int, default=None, required=True,
         choices=[Range(50, 90)], help="Altitudes [km]")
 ARGS = PARSER.parse_args()
